Split DynamoDB and DAX plugins (#200)
Fixes #192
